feat(npm): add majorVersions rule
Add `majorVersions` rule that will deprecate all npm package versions
that are one major version behind the version pointed to by the
`latest` dist-tag if that version is at least X number of months behind
that `latest` version as determined by the number of months passed as
the value of `majorVersions`.

BREAKING CHANGE:

`deprecator` API now takes an object with keys as the name of the rule,
and the value as the value passed to that rule when configuring it.