chore(yarn): add yarn config
Configure the npm registry used by Yarn. This is required when running
the project's unit tests using `yarn test`, and you have a global
`.npmrc` or `.yarnrc` configuration file that configures Yarn to
use a private npm registry. It's not needed if you run
the test scripts directly, or you use `npm test`.